+++ R/DF_to_dt.R
@@ -8,26 +8,36 @@
 #' @param DF \link[S4Vectors]{DataFrame} object.
 #' 
 #' @keywords internal
+#' @importFrom data.table as.data.table
 #' @importFrom methods is
 #' @importFrom Biostrings unstrsplit
 #' @importFrom IRanges CharacterList
 #' @returns VCF data in data.table format.
 DF_to_dt <- function(DF){ 
-    data.table::data.table(
-        mapply(DF, 
-               FUN=function(s){ 
-                   # s <- DF[["REF"]]
-                   # s <- DF[["ALT"]]
-                   # s <- DF[[1]]
-                   if(methods::is(s,"DNAStringSet") ){
-                       s <- as.character(s)
-                   } else if(methods::is(s,"DNAStringSetList")){
-                       s <- IRanges::CharacterList(s)
-                       s <- Biostrings::unstrsplit(s, sep=",")
-                   } else {
-                       s <- as.vector(s)
-                   }
-                   return(s)
-               })
-    )
+    m <- mapply(DF, 
+                FUN=function(s){ 
+                    # s <- DF[["REF"]]
+                    # s <- DF[["ALT"]]
+                    # s <- DF[[1]]
+                    if(methods::is(s,"DNAStringSet") ){
+                        s <- as.character(s)
+                    } else if(methods::is(s,"DNAStringSetList")){
+                        s <- IRanges::CharacterList(s)
+                        s <- Biostrings::unstrsplit(s, sep=",")
+                    } else if(methods::is(s,"NumericList")){
+                        s <- as.numeric(s)
+                    } else if(methods::is(s,"list")){
+                        s <- unlist(s)
+                    } else {
+                        s <- as.vector(s)
+                    }
+                    #### Check if empty ####
+                    if(sum(!is.na(s))==0){
+                        return(NULL)
+                    } else {
+                        return(s)
+                    } 
+                })
+    m <- m[!unlist(lapply(m,is.null))]
+    data.table::as.data.table(m)
 }