+++ R/select_vcf_fields.R
@@ -13,40 +13,43 @@
 #' @importFrom GenomicRanges GRanges
 #' @importFrom methods is
 select_vcf_fields <- function(path,
-                              sampled_rows=1e7,
+                              sampled_rows=1e3,
                               which=NULL,
                               samples=NULL,
                               nThread=1,
                               verbose=TRUE){
     #### Read header ####
     ## Read the first n rows to determine which columns are useful.
-    messager("Finding empty VCF columns based on first",sampled_rows,"rows.",
+    messager("Finding empty VCF columns based on first",
+             formatC(sampled_rows,big.mark = ","),"rows.",
              v=verbose)  
     #### Get header to extract first chromosome name ####
-    header <- VariantAnnotation::scanVcfHeader(file = path)
-    # #### Construct ScanVcfParam object ####
-    param <- VariantAnnotation::ScanVcfParam(
-        which = GenomicRanges::GRanges(
-            paste0(header@reference[[1]],":1-",as.integer(sampled_rows))
-        )
-    )
-    vcf_top <- VariantAnnotation::readVcf(file = path,
-                                          param = param)
+    # header <- VariantAnnotation::scanVcfHeader(file = path)
+    # # #### Construct ScanVcfParam object ####
+    # param <- VariantAnnotation::ScanVcfParam(
+    #     which = GenomicRanges::GRanges(
+    #         paste0(header@reference[[1]],":1-",as.integer(sampled_rows))
+    #     )
+    # )
+    vcf_file <- VariantAnnotation::VcfFile(file = path,
+                                           index = paste0(path,".tbi"),
+                                           yieldSize=sampled_rows) 
+    vcf_top <- VariantAnnotation::readVcf(file = vcf_file)
+    fields <- VariantAnnotation::vcfFields(x = vcf_top)
     # #### Warnings generated bc 0 rows are present ####
-    df <- suppressWarnings(
-        vcf2df(vcf = vcf_top,
-               add_sample_names = FALSE)
-    )
+    df <-  vcf2df(vcf = vcf_top,
+                  add_sample_names = FALSE,
+                  verbose = verbose)
+    remove(vcf_top)
     #### Check n rows returned ####
     if(nrow(df)==0) {
         stop("Query returned no rows. Increase sampled_rows.")
     }
     #### Check for empty cols #####
-    empty_cols <- check_empty_cols(sumstats_dt = df) 
-    fields <- VariantAnnotation::vcfFields(x = path)
-    for(x in names(fields)){
+    remove_empty_cols(sumstats_dt = df)
+    for(x in names(fields)[names(fields)!="samples"]){
         fields[[x]] <- fields[[x]][
-            !toupper(fields[[x]]) %in% toupper(names(empty_cols))
+            fields[[x]] %in% names(df)
             ]
     } 
     #### Select samples ####